requirements.txt
# Core dependencies
streamlit>=1.28.0
numpy>=1.24.3
pandas>=2.0.3
plotly>=5.18.0
scipy>=1.11.1
scikit-learn>=1.3.0

# Code Editor and Syntax Highlighting
streamlit-ace>=0.1.1
pygments>=2.15.1

# Visualization
plotly>=5.18.0
matplotlib>=3.7.1

# AI and LLM Integration
openai>=1.3.0
anthropic>=0.5.0
python-dotenv>=1.0.0

# Audio and Video
opencv-python>=4.8.0
PyAudio>=0.2.13
SpeechRecognition>=3.10.0
moviepy>=1.0.3

# UI Components
streamlit-camera-input-live>=0.2.0
streamlit-drawable-canvas>=0.9.2

# Quantum Computing
qiskit-terra>=0.24.1
qiskit-aer>=0.12.0
qiskit-ibm-runtime>=0.11.0

# Development Tools
pytest>=7.4.0
black>=23.3.0
flake8>=6.0.0

